Require minimum ETH value passed on delegation
To protect the delegation process from an attacks when an adversary
wants to block operator's account with a fake delegation before the real
owner delegates. The protection is a requirement to provide some ETH
with a call. The complete value is deposited in the contract as
unbonded value that can be used later for bonding by the operator.
This commit contains some stub value definition that should be changed
to proper value for mainnet.